Added reports and functionality
Created print_host_certificates_expiring_in_days report. Allowing you to specify a number of days, and it will check for host certificates that expire in that many days or less. Also added function to get a certificate expiration date for a host by name...just in case I need it,.